+++ src/Bootstrap.php
@@ -11,6 +11,7 @@ use PressbooksMultiInstitution\Controllers\AssignBooksController;
 use PressbooksMultiInstitution\Controllers\AssignUsersController;
 use PressbooksMultiInstitution\Controllers\InstitutionsController;
 use PressbooksMultiInstitution\Services\InstitutionStatsService;
+use PressbooksMultiInstitution\Services\MenuManager;
 use PressbooksMultiInstitution\Services\PermissionsManager;
 use PressbooksMultiInstitution\Views\BookList;
 use PressbooksMultiInstitution\Views\UserList;
@@ -112,9 +113,9 @@ final class Bootstrap
         }
         add_action('user_register', fn (int $id) => app(AssignUserToInstitution::class)->handle($id));
         add_action('pb_new_blog', fn () => app(AssignBookToInstitution::class)->handle());
-        add_action('network_admin_menu', fn () => app(PermissionsManager::class)->handleMenus(), 1000);
-        add_action('admin_menu', fn () => app(PermissionsManager::class)->handleMenus(), 1000);
-        add_action('init', fn () => app(PermissionsManager::class)->setupInstitutionalFilters());
+        add_action('network_admin_menu', fn () => app(MenuManager::class)->handleMenus(), 1000);
+        add_action('admin_menu', fn () => app(MenuManager::class)->handleMenus(), 1000);
+        add_action('init', fn () => app(PermissionsManager::class)->setupFilters());
         add_action('pb_institutional_after_save', [PermissionsManager::class, 'syncRestrictedUsers'], 10, 2);
         add_action('pb_institutional_after_delete', [PermissionsManager::class, 'syncRestrictedUsers'], 10, 2);
         add_action(
